commands: check if west is up to date before running
By default, check if west is up to date before running a new command,
emitting a warning if it is not.

This is not appropriate in all situations, e.g. when running the
commands selfupdate and update without -x. To allow for this, add a
new check_up_to_date kwarg to the WestCommand constructor which
disables this check.

Use this new kwarg appropriately from the selfupdate and update
implementations.

Signed-off-by: Marti Bolivar <marti@foundries.io>