project.py: post-refactor clean up of logging calls
While reviewing logging calls during the most recent extraction of
project.py code into manifest.py, it became clear that the colorized
and "=== '-prefixed _banner() routine was being called way too often.

Clean this up by adding a new _msg() routine that is prefixed with
"--- " and not colorized. This still lets the output stand out from
git output, but doesn't clutter the output.

Conversely, some _banner() calls were missing, such as in the Diff
class.

While we are here, normalize the printlines to use {name_and_path} as
a prefix for most _msg calls related to a project, for consistency.

Signed-off-by: Marti Bolivar <marti@foundries.io>