manifest.py: remove 'rename' support from import: map
This feature was originally added when we had a design where imported
projects could partially override the values from parent manifests.

Now that we have a different design where name collisions in imported
manifests cause the submanifest projects to simply be ignored, we
don't really have a clear use case for this feature and should
therefore simply remove it. It is easy to revert this commit if we
decide we really need it.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>